fwutil/azure_interface/requirements.txt
azure-core==1.26.4
azure-storage-blob==12.16.0